requirements.txt
streamlit==1.32.0
openai==1.55.3
supabase==2.11.0
python-dotenv==1.0.1